+++ Backend/src/main/java/com/projeto/biblioteca/repository/EmprestimoRepository.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 import java.util.List;
 
 @Repository
-public interface EmprestimoRepository extends JpaRepository<Emprestimo, Long> {
+public interface EmprestimoRepository extends JpaRepository<Emprestimo, Integer> {
     List<Emprestimo> findByClienteId(Long clienteId);
     List<Emprestimo> findByLivroId(Long livroId);
     List<Emprestimo> findByDataRetiradaBetween(LocalDate inicio, LocalDate fim);
